==> rating_prediction/code/00_select_long_reviews.ipynb <==
import pandas as pd
import polars as pl
import numpy as np
import seaborn as sns

import os
import pickle
from itertools import chain

import gc

with open('/Users/tjstarak/Documents/GitHub/thesis/universe/reviews/completed/reviews_0_41677_pass1.pkl', 'rb') as f:
    place_reviews = pickle.load(f)

reviews = [r for v in place_reviews.values() for r in v['reviews'] if len(r['review_text']) > 0]
long_reviews = [r for r in reviews if len(r['review_text']) >= 500]
del reviews

long_reviews[:10]

sorted_reviews = [dict(sorted(r.items())) for r in long_reviews]

sorted_reviews[:10]

reviews_df = pl.from_dicts(
    sorted_reviews, 
    schema={
        'age': pl.Utf8, 
        'author': pl.Utf8, 
        'author_review_count': pl.Utf8, 
        'rating': pl.Utf8,
        'review_keywords': pl.Object,
        'review_text': pl.Utf8
    }
)

reviews_df.write_csv('long_reviews.csv')

lens = [len(r['review_text']) for r in reviews]
lens = np.array(lens)

np.sum(lens > 500)
