==> rating_prediction/code/00_tokenization.ipynb <==
# - clean data
# -- remove punctuation
# -- remove stopwords (though this is not necessary for NB?)
# -- tokenize
# -- create mega-document
# -- train NB classifier with cross-validation

import pandas as pd
import spacy

reviews = pd.read_parquet('../data/reviews_sample_proc.parquet')[['original', 'translated', 'rating', 'partition', 'original_proc', 'original_proc_no_stop', 'translated_proc', 'translated_proc_no_stop']]
# nlp = spacy.load('pl_core_news_lg')
nlp = spacy.load('en_core_web_lg')

parsed = list(nlp.pipe(reviews['translated']))

reviews_tokenized = [[token.lemma_ for token in review if not token.is_punct] for review in parsed]
reviews_tokenized_no_stop = [[token.lemma_ for token in review if (not token.is_punct) and (not token.is_stop)] for review in parsed]

reviews['translated_proc'] = pd.Series(reviews_tokenized)
reviews['translated_proc_no_stop'] = pd.Series(reviews_tokenized_no_stop)

reviews.columns

reviews['original_proc'] = reviews['original_proc'].apply(lambda x: ' '.join(x.tolist()).lower())
reviews['original_proc_no_stop'] = reviews['original_proc_no_stop'].apply(lambda x: ' '.join(x.tolist()).lower())
reviews['translated_proc'] = reviews['translated_proc'].apply(lambda x: ' '.join(x.tolist()).lower())
reviews['translated_proc_no_stop'] = reviews['translated_proc_no_stop'].apply(lambda x: ' '.join(x.tolist()).lower())

reviews.to_parquet('../data/reviews_sample_proc.parquet')
