==> rating_prediction/code/02_xgb.ipynb <==
import pandas as pd
import numpy as np
import xgboost as xgb

from sklearn.pipeline import make_pipeline
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.model_selection import StratifiedKFold
from sklearn import metrics

from mlxtend.preprocessing import DenseTransformer

import optuna
from optuna.samplers import TPESampler

reviews = pd.read_parquet('/kaggle/input/restaurant-reviews/reviews_sample_proc.parquet')
train = reviews.query('partition == "train"')
test = reviews.query('partition == "test"')

scope = 'original'
X_train = train[scope].to_numpy()
y_train = train['rating'].to_numpy() - 1

X_test = test[scope].to_numpy()
y_test = test['rating'].to_numpy() - 1

vectorizer = TfidfVectorizer(min_df=200)
X_train_vec = vectorizer.fit_transform(X_train[:10000])
X_train_vec.shape

#### Example of Bayesian optimization for max tree depth (the rest

%%time

def objective(trial):
    params = {
        'eta': 0.3,
        'max_depth': 8,
        'alpha': trial.suggest_float('alpha', low=0.0, high=0.3, step=0.01),
        'objective': 'multi:softmax',
        'num_class': 5,
        'eval_metric': 'mlogloss',
        'early_stopping_rounds': 5,
        'n_estimators': 200,
        'tree_method': 'gpu_hist',
    }

    cv = StratifiedKFold(n_splits=3, shuffle=True, random_state=123)

    y_preds = []
    y_true = []
    scores = []
    n_rounds = []
    for (train_ix, test_ix) in cv.split(X_train, y_train):
        vectorizer = TfidfVectorizer(min_df=200)
        X_train_vec = vectorizer.fit_transform(X_train[train_ix])
        X_val_vec = vectorizer.transform(X_train[test_ix])
        clf = xgb.XGBClassifier(**params)
        clf.fit(X_train_vec, y_train[train_ix], eval_set=[(X_val_vec, y_train[test_ix])], verbose=False)
        scores.append(clf.best_score)
        n_rounds.append(clf.best_iteration)
        y_preds.extend(clf.predict(X_val_vec))
        y_true.extend(y_train[test_ix])

    score = metrics.f1_score(y_true, y_preds, average='macro')
    print(np.mean(n_rounds))
    
    return score

study = optuna.create_study(sampler=TPESampler(), direction='maximize')
study.optimize(objective, n_trials=20)

cols = ['Iteration number', 'Score'] + list(study.best_params.keys())
results = [study.best_trial.number, study.best_trial.value] + list(study.best_params.values())
results_df = pd.DataFrame([results], columns = cols)
results_df

#### Finding the optimal number of boosting rounds

%%time

params = {
    'eta': 0.1,
    'max_depth': 8,
    'alpha': 0.2,
    'objective': 'multi:softmax',
    'num_class': 5,
    'eval_metric': 'mlogloss',
    'early_stopping_rounds': 10,
    'n_estimators': 1000,
    'tree_method': 'gpu_hist',
}

cv = StratifiedKFold(n_splits=3, shuffle=True, random_state=123)

y_preds = []
y_true = []
scores = []
n_rounds = []
for (train_ix, test_ix) in cv.split(X_train, y_train):
    vectorizer = TfidfVectorizer(min_df=200)
    X_train_vec = vectorizer.fit_transform(X_train[train_ix])
    X_val_vec = vectorizer.transform(X_train[test_ix])
    clf = xgb.XGBClassifier(**params)
    clf.fit(X_train_vec, y_train[train_ix], eval_set=[(X_val_vec, y_train[test_ix])], verbose=False)
    scores.append(clf.best_score)
    n_rounds.append(clf.best_iteration)
    y_preds.extend(clf.predict(X_val_vec))
    y_true.extend(y_train[test_ix])

score = metrics.f1_score(y_true, y_preds, average='macro')
print(np.mean(n_rounds))
print(score)

#### Polish text, raw (punctuation removed)

%%time

scope = 'original'
X_train = train[scope].to_numpy()
y_train = train['rating'].to_numpy() - 1

X_test = test[scope].to_numpy()
y_test = test['rating'].to_numpy() - 1

params = {
    'eta': 0.1,
    'max_depth': 8,
    'alpha': 0.2,
    'objective': 'multi:softmax',
    'num_class': 5,
    'eval_metric': 'mlogloss',
    'n_estimators': 605,
    'tree_method': 'gpu_hist',
}

vectorizer = TfidfVectorizer(min_df=200)
X_train_vec = vectorizer.fit_transform(X_train)
X_test_vec = vectorizer.transform(X_test)

clf = xgb.XGBClassifier(**params)
clf.fit(X_train_vec, y_train, verbose=False)

y_pred = clf.predict(X_test_vec)

print(metrics.classification_report(y_test, y_pred))

#### Polish text, lemmatized

%%time

scope = 'original_proc'
X_train = train[scope].to_numpy()
y_train = train['rating'].to_numpy() - 1

X_test = test[scope].to_numpy()
y_test = test['rating'].to_numpy() - 1

params = {
    'eta': 0.1,
    'max_depth': 8,
    'alpha': 0.2,
    'objective': 'multi:softmax',
    'num_class': 5,
    'eval_metric': 'mlogloss',
    'n_estimators': 605,
    'tree_method': 'gpu_hist',
}

vectorizer = TfidfVectorizer(min_df=200)
X_train_vec = vectorizer.fit_transform(X_train)
X_test_vec = vectorizer.transform(X_test)

clf = xgb.XGBClassifier(**params)
clf.fit(X_train_vec, y_train, verbose=False)

y_pred = clf.predict(X_test_vec)

print(metrics.classification_report(y_test, y_pred))

#### Polish text, lemmatized with stopwords removed

%%time

scope = 'original_proc_no_stop'
X_train = train[scope].to_numpy()
y_train = train['rating'].to_numpy() - 1

X_test = test[scope].to_numpy()
y_test = test['rating'].to_numpy() - 1

params = {
    'eta': 0.1,
    'max_depth': 8,
    'alpha': 0.2,
    'objective': 'multi:softmax',
    'num_class': 5,
    'eval_metric': 'mlogloss',
    'n_estimators': 605,
    'tree_method': 'gpu_hist',
}

vectorizer = TfidfVectorizer(min_df=200)
X_train_vec = vectorizer.fit_transform(X_train)
X_test_vec = vectorizer.transform(X_test)

clf = xgb.XGBClassifier(**params)
clf.fit(X_train_vec, y_train, verbose=False)

y_pred = clf.predict(X_test_vec)

print(metrics.classification_report(y_test, y_pred))

#### English translation, raw (with punctuation removed)

%%time

scope = 'translated'
X_train = train[scope].to_numpy()
y_train = train['rating'].to_numpy() - 1

X_test = test[scope].to_numpy()
y_test = test['rating'].to_numpy() - 1

params = {
    'eta': 0.1,
    'max_depth': 8,
    'alpha': 0.2,
    'objective': 'multi:softmax',
    'num_class': 5,
    'eval_metric': 'mlogloss',
    'n_estimators': 605,
    'tree_method': 'gpu_hist',
}

vectorizer = TfidfVectorizer(min_df=200)
X_train_vec = vectorizer.fit_transform(X_train)
X_test_vec = vectorizer.transform(X_test)

clf = xgb.XGBClassifier(**params)
clf.fit(X_train_vec, y_train, verbose=False)

y_pred = clf.predict(X_test_vec)

print(metrics.classification_report(y_test, y_pred))

#### English text, lemmatized

%%time

scope = 'translated_proc'
X_train = train[scope].to_numpy()
y_train = train['rating'].to_numpy() - 1

X_test = test[scope].to_numpy()
y_test = test['rating'].to_numpy() - 1

params = {
    'eta': 0.1,
    'max_depth': 8,
    'alpha': 0.2,
    'objective': 'multi:softmax',
    'num_class': 5,
    'eval_metric': 'mlogloss',
    'n_estimators': 605,
    'tree_method': 'gpu_hist',
}

vectorizer = TfidfVectorizer(min_df=200)
X_train_vec = vectorizer.fit_transform(X_train)
X_test_vec = vectorizer.transform(X_test)

clf = xgb.XGBClassifier(**params)
clf.fit(X_train_vec, y_train, verbose=False)

y_pred = clf.predict(X_test_vec)

print(metrics.classification_report(y_test, y_pred))

#### English text, lemmatized with stopwords removed

%%time

scope = 'translated_proc_no_stop'
X_train = train[scope].to_numpy()
y_train = train['rating'].to_numpy() - 1

X_test = test[scope].to_numpy()
y_test = test['rating'].to_numpy() - 1

params = {
    'eta': 0.1,
    'max_depth': 8,
    'alpha': 0.2,
    'objective': 'multi:softmax',
    'num_class': 5,
    'eval_metric': 'mlogloss',
    'n_estimators': 605,
    'tree_method': 'gpu_hist',
}

vectorizer = TfidfVectorizer(min_df=200)
X_train_vec = vectorizer.fit_transform(X_train)
X_test_vec = vectorizer.transform(X_test)

clf = xgb.XGBClassifier(**params)
clf.fit(X_train_vec, y_train, verbose=False)

y_pred = clf.predict(X_test_vec)

print(metrics.classification_report(y_test, y_pred))
